fix(build): satisfy strict warnings on Apple Clang and pin vcpkg baseline
Compiler-flag and linker-flag probes (CompilerFlags.cmake, Hardening.cmake) now run check_cxx_compiler_flag with -Werror in CMAKE_REQUIRED_FLAGS so flags that Apple Clang accepts with an 'unused command line argument' warning (e.g. -fno-semantic-interposition, -fstack-clash-protection, -fcf-protection=full) are correctly rejected by the probe and never reach the real compile line.

Bump cmake_minimum_required to 3.30 in both CMakeLists.txt and CMakePresets.json so policy CMP0167 implicitly resolves to NEW (Boost ships its own CMake config; the old FindBoost module emits a deprecation warning otherwise).

vcpkg.json + vcpkg-configuration.json now carry a real builtin-baseline pinned to microsoft/vcpkg HEAD at the time of this work. vcpkg 2025+ rejects builtin registry entries without a baseline.

bench/bench_arena.cpp, bench/bench_bitmap.cpp, and bench/bench_spsc.cpp use explicit static_cast<std::int64_t> around state.SetItemsProcessed inputs to satisfy -Wsign-conversion (state.iterations() is signed; std::size_t multipliers are unsigned). Install rules removed pending a settled engine ABI.

+++ bench/bench_arena.cpp
@@ -30,7 +30,8 @@ void bench_alloc_only(benchmark::State& state) {
         live.clear();
         state.ResumeTiming();
     }
-    state.SetItemsProcessed(state.iterations() * n);
+    state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+                            static_cast<std::int64_t>(n));
 }
 BENCHMARK(bench_alloc_only)->Range(64, 32'768);
 
@@ -47,7 +48,8 @@ void bench_dealloc_only(benchmark::State& state) {
         live.clear();
         benchmark::ClobberMemory();
     }
-    state.SetItemsProcessed(state.iterations() * n);
+    state.SetItemsProcessed(static_cast<std::int64_t>(state.iterations()) *
+                            static_cast<std::int64_t>(n));
 }
 BENCHMARK(bench_dealloc_only)->Range(64, 32'768);
